requirements.txt
# --- Core de la Aplicación y Web ---
gradio>=4.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
prometheus-client>=0.19.0

# --- Librerías de LangChain (Conjunto compatible) ---
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-huggingface>=0.0.3
langchain-elasticsearch>=0.1.0
langchain-postgres>=0.0.14

# --- Bases de Datos Vectoriales y Conectores ---
faiss-cpu>=1.7.4
redis>=5.0.0
elasticsearch>=8.8.0
pgvector>=0.2.0
psycopg[binary]>=3.1.0

# --- IA, Embeddings y Procesamiento de Datos ---
sentence-transformers>=2.2.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# FIX: Agregar huggingface-hub actualizado para evitar warnings
huggingface-hub>=0.20.0
transformers>=4.36.0
tokenizers>=0.15.0

# --- Utilidades Generales ---
pydantic>=2.0.0
httpx>=0.24.0
aiohttp>=3.8.0
markdown>=3.5.0
pdfkit>=1.0.0

# FIX: Para evitar warnings de SSL
urllib3>=2.0.0
certifi>=2023.0.0

lark>=1.1.9